Search harder for VMRUNNER and CHAINLOADER paths
Now that vmrunner is a python package we can use the metadata to find
the script location. If INCLUDEOS_VMRUNNER is not set, this adds support
for finding the location automatically.

If vmrunner is installed with nix, the chainloader will be in a path
listed in the propagatedBuildInputs environment variable. If
INCLUDEOS_CHAINLOADER is not set, search for `/bin/chainloader` in this
path to attempt to find it.